+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/spi/ConverterService.java
@@ -18,7 +18,7 @@ public abstract class ConverterService {
     private final Map<Class<?>, SqlAccessor> cache = new ConcurrentHashMap<>();
 
     public static synchronized ConverterService getInstance() {
-        if (!INSTANCE.isPresent()) {
+        if (!INSTANCE.isPresent() || FrameworkIntegrationService.isReloadRequired(INSTANCE.get().getConverters().keySet())) {
             INSTANCE.set(new DefaultConverters(ServiceFinder.loadServices(ConverterProvider.class)));
         }
 
@@ -32,7 +32,7 @@ public abstract class ConverterService {
         }
         ConverterPair<T,R> converterPair = (ConverterPair<T, R>) getConverters().entrySet()
                 .stream()
-                .filter(el -> ClassChecker.isAssignable(el.getKey(), klass))
+                .filter(el -> el.getKey().equals(klass))
                 .findFirst()
                 .map(Map.Entry::getValue)
                 .orElse(null);