+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/simple/FromSimpleSelected.java
@@ -4,5 +4,6 @@ import io.github.ulisse1996.jaorm.dsl.query.common.trait.WithSubQuerySupport;
 import io.github.ulisse1996.jaorm.dsl.query.simple.trait.*;
 
 public interface FromSimpleSelected extends WithProjectionResult,
-        WithSimpleJoin, WithSimpleOrder, WithSimpleWhere, WithSubQuerySupport, WithSimpleLimit, WithSimpleOffset {
+        WithSimpleJoin, WithSimpleOrder, WithSimpleWhere,
+        WithSubQuerySupport, WithSimpleLimit, WithSimpleOffset, WithSimpleGroup {
 }