+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/simple/intermediate/SimpleGroup.java
@@ -1,0 +1,6 @@
+package io.github.ulisse1996.jaorm.dsl.query.simple.intermediate;
+
+import io.github.ulisse1996.jaorm.dsl.query.simple.trait.*;
+
+public interface SimpleGroup extends WithSimpleGroup, WithSimpleHaving,
+        WithProjectionResult, WithSimpleOffset, WithSimpleLimit {}