+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/simple/intermediate/SimpleSelectedWhere.java
@@ -2,7 +2,8 @@ package io.github.ulisse1996.jaorm.dsl.query.simple.intermediate;
 
 import io.github.ulisse1996.jaorm.dsl.query.simple.trait.WithAndOrWhere;
 import io.github.ulisse1996.jaorm.dsl.query.simple.trait.WithProjectionResult;
+import io.github.ulisse1996.jaorm.dsl.query.simple.trait.WithSimpleGroup;
 import io.github.ulisse1996.jaorm.dsl.query.simple.trait.WithSimpleOrder;
 
-public interface SimpleSelectedWhere extends WithAndOrWhere, WithSimpleOrder, WithProjectionResult {
+public interface SimpleSelectedWhere extends WithAndOrWhere, WithSimpleOrder, WithSimpleGroup, WithProjectionResult {
 }